[UPDATE] Split Step 1 challenges into different files and update RUNME.md file
